Install @vue/test-utils to reduce cdn dependency
@vue/test-utils was installed via cdn
because of 'process.env' issues with Vue
and circular dependency on '@vue/test-utils'.

They're fixed now and the unfixable rollup warnings are ignored.
Local testing without cdn should work now.

Included js-beautify to maintain support for Nodejs 8.
It went EOL this year but I am using it to support Ubuntu 18.04.