Support SSR for table mixin.
Vue SSR does not have 'mounted' lifecycle event.
Its last lifecycle event is 'serverPrefetch' which can be asynchronous.
smart-table uses "setTimeout" to delay execution.
Wrap a timer with a promise to wait for smart-table
to initialize the the displayed table rows.

Test the code on Nodejs by using @vue/server-test-utils.
"zora", test runner, uses asynchronous generators.
Compile the test code and dependencies with rollup and babel
so that it works on Nodejs 8.